# backend/internal/swagger/openapi.yaml
openapi: 3.0.3
info:
  title: LogiFlows API Gateway
  description: Authoritative REST API for LogiFlows Smart Postal & Courier Delivery Management System
  version: 0.3.0-fleet-telematics
servers:
  - url: /api/v1
    description: Current Server
  - url: http://localhost:8080/api/v1
    description: Local Development Server
...
paths:
...
  /serviceability/check:
    post:
      tags:
        - Geospatial Serviceability
      summary: Check Coordinate Serviceability
      description: Evaluates whether a coordinate falls inside any active branch service area polygon using PostGIS ST_Contains
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceabilityCheckRequest'
      responses:
        '200':
          description: Serviceability evaluated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceabilityCheckResponse'
        '400':
          description: Invalid coordinates or payload
    get:
      tags:
        - Geospatial Serviceability
      summary: Check Coordinate Serviceability (Query Parameters)
      parameters:
        - name: tenant_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: lat
          in: query
          required: true
          schema:
            type: number
            example: 12.9716
        - name: lon
          in: query
          required: true
          schema:
            type: number
            example: 77.5946
      responses:
        '200':
          description: Serviceability evaluated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceabilityCheckResponse'
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register User
      description: Registers a new user account with default CUSTOMER role unless specified
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Validation error
        '409':
          description: Email already registered
  /auth/login:
    post:
      tags:
        - Authentication
      summary: User Login
      description: Authenticates user credentials and issues JWT access token + refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid email or password
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Rotate Access & Refresh Tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token:
                  type: string
      responses:
        '200':
          description: Tokens refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid or expired refresh token
  /auth/logout:
    post:
      tags:
        - Authentication
      summary: Revoke Refresh Token
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token:
                  type: string
      responses:
        '200':
          description: Token revoked successfully
  /auth/me:
    get:
      tags:
        - Authentication
      summary: Get Current Profile & Permissions
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Current authenticated profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
        '401':
          description: Missing or invalid token
  /tenants:
    post:
      tags:
        - Tenants
      summary: Create Tenant
      description: Restricted to PLATFORM_ADMIN
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TenantCreateRequest'
      responses:
        '201':
          description: Tenant created successfully
    get:
      tags:
        - Tenants
      summary: List Tenants
      description: Platform Admins list all tenants; tenant users list their own tenant
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of tenants
  /tenants/{tenant_id}:
    get:
      tags:
        - Tenants
      summary: Get Tenant Details
      security:
        - BearerAuth: []
      parameters:
        - name: tenant_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Tenant details
    put:
      tags:
        - Tenants
      summary: Update Tenant
      security:
        - BearerAuth: []
      parameters:
        - name: tenant_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Tenant updated
...
  /admin/system-check:
    get:
      tags:
        - RBAC & Tenant Verification
      summary: Platform Admin System Check
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Platform admin access granted
  /parcels/permission-check:
    get:
      tags:
        - RBAC & Tenant Verification
      summary: Parcel Create Permission Check
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User has parcel.create permission
...
  /vehicles:
    get:
      tags:
        - Fleet Management
      summary: List Tenant Fleet Vehicles
      security:
        - BearerAuth: []
      parameters:
        - name: branch_id
          in: query
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            type: string
            enum: [AVAILABLE, ON_ROUTE, MAINTENANCE, DECOMMISSIONED]
        - name: type
          in: query
          schema:
            type: string
            enum: [VAN, MOTORCYCLE, TRUCK, ELECTRIC_VAN, BICYCLE]
      responses:
        '200':
          description: Vehicles listed successfully
    post:
      tags:
        - Fleet Management
      summary: Register Fleet Vehicle
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VehicleCreate'
      responses:
        '201':
          description: Vehicle created successfully
  /vehicles/{vehicle_id}:
    get:
      tags:
        - Fleet Management
      summary: Get Vehicle Details
      security:
        - BearerAuth: []
      parameters:
        - name: vehicle_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Vehicle details returned
    put:
      tags:
        - Fleet Management
      summary: Update Vehicle Details
      security:
        - BearerAuth: []
      parameters:
        - name: vehicle_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Vehicle updated successfully
    delete:
      tags:
        - Fleet Management
      summary: Delete Vehicle
      security:
        - BearerAuth: []
      parameters:
        - name: vehicle_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Vehicle deleted
  /vehicles/{vehicle_id}/telematics:
    post:
      tags:
        - Vehicle Telematics
      summary: Ingest GPS Telematics Ping
      security:
        - BearerAuth: []
      parameters:
        - name: vehicle_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VehicleTelematicsPing'
      responses:
        '201':
          description: Telematics recorded
  /vehicles/{vehicle_id}/telematics/latest:
    get:
      tags:
        - Vehicle Telematics
      summary: Get Latest Vehicle Telematics
      security:
        - BearerAuth: []
      parameters:
        - name: vehicle_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Latest telemetry returned
  /vehicles/{vehicle_id}/telematics/history:
    get:
      tags:
        - Vehicle Telematics
      summary: Get Telematics Breadcrumbs History
      security:
        - BearerAuth: []
      parameters:
        - name: vehicle_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
      responses:
        '200':
          description: Telematics history returned
...
  /vehicles/{vehicle_id}/maintenance:
    get:
      tags:
        - Fleet Maintenance
      summary: List Vehicle Maintenance Records
      security:
        - BearerAuth: []
      parameters:
        - name: vehicle_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Maintenance records listed
    post:
      tags:
        - Fleet Maintenance
      summary: Log Maintenance Intervention
      security:
        - BearerAuth: []
      parameters:
        - name: vehicle_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '201':
          description: Maintenance logged
  /fleet/maintenance/upcoming:
    get:
      tags:
        - Fleet Maintenance
      summary: List Vehicles Due for Maintenance
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Upcoming maintenance list returned
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ServiceabilityCheckRequest:
      type: object
      required: [tenant_id, latitude, longitude]
      properties:
        tenant_id:
          type: string
          format: uuid
        latitude:
          type: number
          example: 12.9716
        longitude:
          type: number
          example: 77.5946
        address:
          type: string
          example: "MG Road, Bengaluru"
    ServiceabilityCheckResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            is_serviceable:
              type: boolean
            status:
              type: string
              example: SERVICEABLE
            message:
              type: string
            matched_branch:
              $ref: '#/components/schemas/BranchSummary'
            nearest_branch:
              $ref: '#/components/schemas/BranchSummary'
            distance_meters:
              type: number
            estimated_eta:
              type: string
    BranchSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
          example: BLR-HUB-01
        name:
          type: string
          example: Bengaluru Central Sorting Hub
        branch_type:
          type: string
          enum: [SORTING_HUB, DISTRIBUTION_CENTER, LOCAL_OFFICE]
        address:
          type: string
        city:
          type: string
        contact_phone:
          type: string
        status:
          type: string
          enum: [ACTIVE, INACTIVE, MAINTENANCE]
    RegisterRequest:
      type: object
      required: [email, password, first_name, last_name]
      properties:
        email:
          type: string
        password:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [PLATFORM_ADMIN, TENANT, TENANT_ADMIN, EMPLOYEE, CUSTOMER]
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            user:
              $ref: '#/components/schemas/User'
            tokens:
              $ref: '#/components/schemas/TokenPair'
    MeResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            user:
              $ref: '#/components/schemas/User'
            permissions:
              type: array
              items:
                type: string
    TenantCreateRequest:
      type: object
      required: [name, code]
      properties:
        name:
          type: string
        code:
          type: string
        slug:
          type: string
        tier:
          type: string
          example: ENTERPRISE
        quota_parcels_per_day:
          type: integer
          example: 50000
        quota_branches:
          type: integer
          example: 100
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [PLATFORM_ADMIN, TENANT, TENANT_ADMIN, EMPLOYEE, CUSTOMER]
        is_active:
          type: boolean
        tenant_id:
          type: string
          format: uuid
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    TokenPair:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        expires_at:
          type: string
          format: date-time
        token_type:
          type: string
          example: Bearer
...
    VehicleCreate:
      type: object
      required:
        - license_plate
        - make
        - model
        - year
        - vehicle_type
        - fuel_type
        - capacity_kg
        - capacity_volume_m3
        - max_parcels
      properties:
        assigned_branch_id:
          type: string
          format: uuid
        assigned_driver_id:
          type: string
          format: uuid
        license_plate:
          type: string
          example: KA-01-EQ-4091
        vin:
          type: string
          example: MAT90812903X77821
        make:
          type: string
          example: Tata
        model:
          type: string
          example: Ace EV
        year:
          type: integer
          example: 2024
        vehicle_type:
          type: string
          enum: [VAN, MOTORCYCLE, TRUCK, ELECTRIC_VAN, BICYCLE]
        fuel_type:
          type: string
          enum: [DIESEL, PETROL, ELECTRIC, HYBRID]
        capacity_kg:
          type: number
          example: 1000
        capacity_volume_m3:
          type: number
          example: 5.8
        max_parcels:
          type: integer
          example: 80
        current_mileage_km:
          type: number
          example: 1420.5
        battery_or_fuel_level_percent:
          type: number
          example: 92.5
        status:
          type: string
          enum: [AVAILABLE, ON_ROUTE, MAINTENANCE, DECOMMISSIONED]
        metadata:
          type: object
    VehicleTelematicsPing:
      type: object
      required:
        - latitude
        - longitude
        - speed_kmh
        - heading_degrees
        - battery_or_fuel_percent
        - odometer_km
      properties:
        latitude:
          type: number
          example: 12.9716
        longitude:
          type: number
          example: 77.5946
        speed_kmh:
          type: number
          example: 38.5
        heading_degrees:
          type: number
          example: 142.0
        battery_or_fuel_percent:
          type: number
          example: 88.0
        odometer_km:
          type: number
          example: 1450.2
        recorded_at:
          type: string
          format: date-time
...